requirements.txt
PyQt6>=6.5
pyusb>=1.2
